change in home side calculations
change in the way the home side is found from goals in a season to shot in a game. Some teams might play singular game in another venue (winter classic, europe/china games) therefore taking the season average is prone to errors. Would need to check if this is causing issue due to teams not shooting during a period but it should be a rare event.