Update tidy_dataframe_builder.py
better implementation to find shooter + goalie. This update makes sure we get the shooter even if he's not the first player listed in the event and the goalie even if he's not the last player in the event. If no goalie is found, we will have a NaN value in place of the goalie name.